Move polyamorous from gem to a module.
