+++ lib/polyamorous/activerecord_6.0_ruby_2/join_dependency.rb
@@ -1,5 +1,4 @@
 # active_record_6.0_ruby_2/join_dependency.rb
-
 module Polyamorous
   module JoinDependencyExtensions
     # Replaces ActiveRecord::Associations::JoinDependency#build
