+++ lib/polyamorous/activerecord_6.0_ruby_2/reflection.rb
@@ -1,2 +1,1 @@
-# active_record_6.0_ruby_2/reflection.rb
 require 'polyamorous/activerecord_5.2_ruby_2/reflection'
