+++ lib/polyamorous/activerecord_6.1_ruby_2/join_association.rb
@@ -1,2 +1,1 @@
-# active_record_6.1_ruby_2/join_association
 require 'polyamorous/activerecord_6.0_ruby_2/join_association'
