+++ lib/polyamorous/join.rb
