+++ lib/polyamorous/polyamorous.rb
