+++ lib/polyamorous/swapping_reflection_class.rb
