+++ lib/ransack.rb
@@ -1,7 +1,7 @@
 require 'active_support/core_ext'
 require 'ransack/configuration'
 require 'ransack/adapters'
-require 'polyamorous'
+require 'polyamorous/polyamorous.rb'
 
 Ransack::Adapters.object_mapper.require_constants
