+++ spec/polyamorous/join_association_spec.rb
