+++ spec/polyamorous/join_dependency_spec.rb
