+++ spec/polyamorous/join_spec.rb
